+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
-* Shield shaped absorption hearts
+## Upcoming
+* Regenerating absorption no longer uses vanilla absorption
+  * Instead, a new shield like icons will appear above the XP bar
 
 ## 3.1.13
 * I hate client-server stuff